chore: Add comments to prometheus receiver config generation and improve var names
